Add conversion of internal graph back to world
